Выносит логику создания массива описаний фото в отдельный модуль
